all: bootstrap using `release-1.5` tree
Previously, the version of LFS that we used to boostrap used the old naming
convention: 'github.com/github/git-lfs". The Dockerfiles in this repository
assume the latest naming convention "github.com/git-lfs/git-lfs".

This commit bumps the bootstrapped version to `release-1.5` so we can start
using the new naming conventions throughout.