centos: don't build Git LFS during setup
We build with the Docker containers primarily on Actions these days,
where the containers are created new each time.  Building Git LFS during
the setup stage means that we must run the entire build and testsuite
for a container which will only be used once, which is very expensive.

Remove the code which builds during the setup stage of the Docker
containers to make the process faster.