build_dockers: pass in hash of Go archive
If someone wants to build the images using a different version of Go,
they can set the environment variable to adjust it accordingly.
However, we also need to allow people to specify the SHA-256 hash of the
appropriate archive, since different versions will necessarily have
different hashes and we'd otherwise have a mismatch.