build_dockers: add a -j option to parallelize
Currently, building our Docker containers sequentially takes about 14
minutes from a fresh image on a reasonably fast connection.  This
expense must be incurred every time we run a new CI job.

To counter this, let's teach build_dockers.bsh a command to run things
in parallel using GNU parallel.  The command works very similarly to
xargs, and we can continue to default to running things in serial if the
user doesn't specify the -j option. In order to make this possible, move
the actual build process for a single image into a separate script so
it's easy to run from parallel.

Note that before this change, the job took 14 minutes on a MacBook Pro
2018 running Linux; afterwards, it takes 6.  The improvements in CI will
be less noticeable since we have fewer processors, but should still be
nontrivial.