debian: install packages for building with cgo on i686
In CI, we build our Debian packages with cgo enabled.  In order to do so
successfully, we need a compiler for i686 and the shared libraries from
the Debian i386 architecture in order so that dpkg-shlibdeps can specify
the proper dependencies.  Let's install these packages in order to make
it possible to build our i386 packages in the same container as amd64.