CentOS: run yum upgrade before continuing
Due to the expiration of DST Root CA X3, CentOS 7 needs an upgrade to
its ca-certificates package to remove this CA so that it can resolve
sites using Let's Encrypt certificates, such as kernel.org, since its
version of OpenSSL is too old to automatically find a trusted path.
Since we use kernel.org to get the Git tarball to replace the obsolete
version shipped in CentOS 7, this is required for our Docker builds to
work there.

Let's make sure our CentOS containers run yum upgrade before installing
other packages so that we have a fully functional, secure system as a
baseline.