Update Go to 1.17.6
Bump the version and SHA-256 hashes for Go 1.17.6.  Don't bump the
individual Dockerfiles, since now that we have two architectures, it's
impossible to specify a correct hash for both of them and the user will
practically always have to use the script we're updating.  This is what
CI does, in any event.