Add a script to bump hashes of files
Right now, updating the version and hashes in build_dockers.bsh and in
the Dockerfiles is a manual hassle involving editing multiple files.  In
addition, if the maintainer performing this task makes a typo (which
this author has previously done), the CI system is broken until they
notice.

Let's automate this process and make our lives easier by using the JSON
API for Go to download the hashes and update the files automatically.
This ensures that the maintainer does not make any typos and that the
checksums are always correct.

Note that this is currently a very rough program without any niceties
like `--help` output or error checking.  However, that's okay, since
it's only intended to be run by maintainers, and we can always add that
in the future.