Debian: vendor dependencies before build
We're going to stop vendoring dependencies in the Git LFS repository.
However, Debian configures the Go binary not to download data from the
Internet during the package build, which obviously means that
dependencies are not satisfied.

To avoid a problem, run `make vendor` to download appropriate
dependencies before the build so that the build process continues to
work.