centos: remove DOCKER_LFS_BUILD_VERSION env var
In commit 330be9d601b234dbc75b4fe68c3a76fd4a7d49e2 of PR #30 we
removed the directives from the CentOS Dockerfiles which compiled a
version of Git LFS using a stable release version during the container
image setup/build stage, as was originally done to save compilation
time later when the containers were run.  This was no longer valuable
as it increased the overall runtime in GitHub Actions, where we build
containers but only use them once, so we removed the directives.

However, we did not remove the supporting references to the
DOCKER_LFS_BUILD_VERSION environment variable which could be used
to set the specific version of the "bootstrapped" Git LFS build,
so we remove them now along with the relevant documentation.