centos_script.bsh: only export git-lfs RPMs
We use an rsync command to copy our built RPM packages into the output
directory specified by the REPO_DIR variable, which defaults to /repo.

At present, this command copies all the RPM files, which includes the
ones we built for the asciidoctor Ruby gem (and prior to Git LFS v3.3.0,
the ones we built for the hpricot, mustache, rdiscount, and ronn Ruby
gems).  As a result we publish RPMs such as
rubygem-asciidoctor-2.0.17-1.el9.src.rpm and
rubygem-asciidoctor-2.0.17-1.el9.noarch.rpm in our Packagecloud account.

To exclude these RPMs so we do not upload them to our Packagecloud
account, we add --include and --exclude options to the rsync command
which limits copying to files with a "git-lfs" prefix.