centos_script.bsh: run chown on given repo dir
In commit 27eed9bf3775a8508cab436bb35e30d047474885 of PR #3 the
debian_script.bsh script was updated to replace several references
to a fixed /repo output directory with the REPO_DIR variable, including
in the final chown command.  However, the same change was not made
in the centos_script.bsh script, so we make it now.