README.md: update and simplify text
Much of the initial text of the README file no longer applies, as we do
not publish Docker images to the Docker Hub or any other registry.
Therefore can we simplify the README considerably, and update the
remaining portions that are still relevant to reflect our current practices,
project locations, and terminology.