*.Dockerfile: force use of local Go toolchain
Since commit 99652c393fcbd8b3b1da33eacd7230da8d00cc4f in PR #31 our
Dockerfiles verify the SHA-256 hash of the Go release package they
download and then use to build Git LFS.

As described in git-lfs/git-lfs#5477, we want to avoid any deviation
from the use of this verified Go release package, and so before we
upgrade to Go 1.21, we need to ensure the new GOTOOLCHAIN environment
variable is set to "local" before any Go commands are executed.
This will prevent the behaviour, added in Go 1.21, whereby Go may
download and use a newer version of itself if it detects certain
conditions, such as a Go module dependency which requires a newer
version of Go than the one installed locally.  While we expect to be
able to avoid such conditions in general, defining GOTOOLCHAIN=local
will guarantee that if one occurs, Go will stop with an error rather
than proceed to download and run an unverified version of itself.