update-hashes: avoid making Dockerfiles executable
The update-hashes script updates the version number and SHA-256
package hashes for new distributions of Go, in both our build_dockers.bsh
script and our Dockerfiles.  At the moment it sets the file permissions
mode of all the files it modifies to make them executable, but only
the build_dockers.bsh script needs those file permissions, so we adjust
the update-hashes script to make the Dockerfiles non-executable.