centos_7.Dockerfile: install from vault.centos.org
CentOS 7 reached its EOL in June 2024, and so the mirror.centos.org
and mirrorlist.centos.org domains no longer resolve.

As we hope to be able to support SUSE Linux Enterprise Server 12.5
until its long-term support ends in October 2027, or at least until
its regular support ends in October 2024, we want to continue building
Docker images with CentOS 7 if possible until then.

For that reason, we update our CentOS 7 Dockerfile to rewrite the
files under /etc/yum.repos.d to point to vault.centos.org instead.
We need to comment the mirrorlist entries and uncomment the baseurl
ones, and we also have to repeat this process for the new files
added after the centos-release-scl package is installed.