build_dockers.bsh,build_one: update comments
In a previous commit in this PR we removed the Dockerfile we used to
build RPM packages for the Red Hat Enterprise Linux (RHEL) 7, CentOS 7,
and SUSE Linux Enterprise Server (SLES) 12 platforms.

As well, in 2019 we removed the Dockerfile we used to build Debian
packages for the Debian 7 platform in commit
5e7bf33b5b2ce784d2060788912a55c448c021ee of PR #26.

Several of our shell scripts still reference these platforms in their
comments, however, so we update those scripts now to provide more
up-to-date usage examples.