debian_10.Dockerfile: drop Debian 10 platform
As announced with the v3.6.0 release of the Git LFS client, and as
noted in commit git-lfs/git-lfs@23fbfe8725553ad58ab8a50dad6fa19c6a602dc7
of PR git-lfs/git-lfs#5911, since all the distribution versions based on
Debian 10 ("buster") for which we previously built Debian packages have
now reached the end of their standard LTS (Long-Term Support) periods,
future releases of Git LFS will no longer provide support for this
platform or build packages for it.

Because we anticipate making a v3.7.0 release of the Git LFS client in
the near future, we now remove the Dockerfile we used to build packages
for the Debian 10 platform.