rocky_10.Dockerfile: add RHEL/Rocky 10 platform
Rocky Linux 10.0 was recently released, following the release last month
of RHEL (Red Hat Enterprise Linux) 10.0, and because we anticipate making
a v3.7.0 release of the Git LFS client in the near future, we now add a
Dockerfile we can use to build packages for the RHEL/Rocky Linux 10
platform.

Note that except for the name of the base image in the FROM statement,
this new Dockerfile is identical to the one for the RHEL/Rocky Linux 9
platform.