upgrade Go to 1.24.4
Because we anticipate making a v3.7.0 release of the Git LFS client in
the near future, we first run our "update-hashes" script to upgrade our
Docker builds to the latest patch version of Go, which is version 1.24.4
at present.

This upgrade should help address the concerns from git-lfs/git-lfs#6064
regarding a recently reported security vulnerability in the "crypto/x509"
Go module.  While this issue does not affect the Git LFS client, some
security scanners will continue to report the Git LFS client as vulnerable
until we release a new version built with the latest version of Go.