install Asciidoctor as RPM package
In PR git-lfs/git-lfs#5054 we updated the source files of the manual
pages for the Git LFS client from the Ronn format to the AsciiDoc
format, which necessitated changing our Linux package build processes
to use the Asciidoctor Ruby gem to generate manual pages in Roff and
HTML formats.

At the time, an RPM package for a 2.0.x version of the Asciidoctor
Ruby gem was not easily available for all the platforms based on Red Hat
Enterprise Linux (RHEL) that we supported, particularly RHEL/CentOS 7,
so in commit git-lfs/git-lfs@db9a82132a2bb066876d8ddf06c5255da2f199a4 of
PR git-lfs/git-lfs#5054 we updated the "rpm/build_rpms.bsh" script in
our main project's repository to build and install a custom RPM
package with Asciidoctor v2.0.17.  In the same commit we also defined
a SPEC file for this custom rubygem-asciidoctor RPM package in the
"rpm/SPECS" directory in our main project.

However, as noted in commit cfde13039f369d3964ef1d62f8c248b40ce6b680
of PR #71, all the distribution versions based on RHEL 7, CentOS 7,
and SUSE Linux Enterprise Server (SLES) 12 for which we previously
built RPM packages have now reached the end of their standard LTS
(Long-Term Support) periods, and so future releases of the Git LFS
client will no longer build packages for them.

This change means we will not need to build and install a custom RPM
package for Asciidoctor any more, because there are rubygem-asciidoctor
packages available in the Extra Packages for Enterprise Linux (EPEL)
collection which provide Asciidoctor v2.0.15 or higher for each of the
RHEL/CentOS 8, RHEL/Rocky Linux 9, and RHEL/Rocky Linux 10 platforms.

We therefore update our Dockerfiles for these platforms to install the
rubygem-asciidoctor package.  As this package is available from the EPEL
collection, we first need to enable the PowerTools (for RHEL/CentOS 7)
or CodeReady Linux Builder (CRB) repository, and then install the EPEL
package, before installing the rubygem-asciidoctor package.  See,
for reference:

  https://docs.fedoraproject.org/en-US/epel/getting-started/
  https://packages.fedoraproject.org/pkgs/rubygem-asciidoctor/rubygem-asciidoctor/

Once this PR is merged, we can then update the "rpm/build_rpms.bsh" script
in our main project so it skips trying to build or install a custom
RPM package for Asciidoctor, and we can remove the corresponding SPEC
file for that package as well.